+++ packages/sdk/tests/test-utils.ts
@@ -17,7 +17,7 @@ const TEST_CHAIN = {
   smsDebugURL: DRONE ? 'http://sms-debug:13300' : 'http://127.0.0.1:13301',
   resultProxyURL: DRONE
     ? 'http://result-proxy:13200'
-    : 'http://127.0.0.1:13200',
+    : 'http://127.0.0.1:13200', // TODO remove
   iexecGatewayURL: DRONE ? 'http://market-api:3000' : 'http://127.0.0.1:3000',
   provider: new JsonRpcProvider(
     DRONE ? 'http://bellecour-fork:8545' : 'http://localhost:8545'