+++ packages/sharing-smart-contract/hardhat.config.cjs
@@ -2,6 +2,7 @@ require('@nomicfoundation/hardhat-foundry');
 require('@nomicfoundation/hardhat-toolbox');
 require('@openzeppelin/hardhat-upgrades');
 require('hardhat-contract-sizer');
+require('@openzeppelin/hardhat-upgrades');
 require('dotenv').config();
 
 const { WALLET_PRIVATE_KEY } = process.env;
@@ -17,7 +18,6 @@ const bellecourBase = {
  */
 module.exports = {
   // run `npx hardhat node` to start the forked bellecour node "local-bellecour-fork"
-  defaultNetwork: 'local-bellecour-fork',
   networks: {
     hardhat: {
       ...bellecourBase,
@@ -30,10 +30,6 @@ module.exports = {
     'local-bellecour-fork': {
       ...bellecourBase,
       url: 'http://127.0.0.1:8545',
-    },
-    'ci-bellecour-fork': {
-      ...bellecourBase,
-      url: 'http://bellecour-fork:8545',
     },
     bellecour: {
       ...bellecourBase,